Fix goodtimeswithscar not correctly reviving after a fire/posion death
