+++ common/types/game-state.ts
@@ -129,11 +129,9 @@ export type PlayerState = {
 		 *
 		 * This is a great place to add blocked actions for the turn, as it's called before actions are calculated
 		 */
-		onTurnStart: GameHook<(attacks: Array<AttackModel>) => void>
+		onTurnStart: GameHook<() => void>
 		/** Hook called at the end of the turn */
 		onTurnEnd: GameHook<(drawCards: Array<CardT | null>) => void>
-		/** Hook called when the time runs out*/
-		onTurnTimeout: GameHook<(newAttacks: Array<AttackModel>) => void>
 
 		/** Hook called the player flips a coin */
 		onCoinFlip: GameHook<(id: string, coinFlips: Array<CoinFlipT>) => Array<CoinFlipT>>
